Refactor spec for removing passenger from flight
